feat(utils): Support serializing Error objects
Modified the `structuredStrictStringify` function to properly handle and serialize `Error` objects.

Previously, this function would identify `Error` objects as an unserializable type and throw an exception. This change now allows `Error` instances to be serialized into an object that includes their `name`, `message`, and `stack` properties. This provides a more useful representation for debugging and inter-process communication (IPC).