chore: update package versions and dependencies in package.json and pnpm-lock.yaml
This commit updates the version of the package to 2.2.0 and modifies the dependencies in both package.json and pnpm-lock.yaml. It removes several unused dependencies and adds new ones, including 'get-windows', 'mac-windows', and 'pinyin-match', to enhance functionality. Additionally, it adjusts the electron.vite.config.ts to refine the handling of libsql packages, ensuring better compatibility and organization in the build output.